Common is as common is used * Moved transfer status and type code definitions from dto to common * Updated references and Javadoc accordingly
