+++ server/bank-accounts/src/test/java/com/pissouri/service/TransferServiceUTest.java
@@ -20,7 +20,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
-import static com.pissouri.dto.TransferStatusCode.*;
+import static com.pissouri.common.TransferStatusCode.*;
 import static java.time.ZoneOffset.UTC;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
